+++ tests/scraper/scrape_performances_test.py
@@ -254,3 +254,22 @@ def test_unexpected_performances(expected_performances, scraped_performances):
     """The scrape_performances method should return a list of dictionaries that does not contain any unexpected values"""
 
     check_unexpected_items(expected_performances, scraped_performances, ['date', 'horse_url', 'track'])
+
+
+def test_mangled_starting_price(scraper, source_timezone):
+    """The scrape_performances method should handle mangled starting prices (see issue #42)"""
+
+    horse = {
+        'url':  'https://www.punters.com.au/horses/Husna_219457/'
+    }
+
+    performances = scraper.scrape_performances(horse)
+
+    found_performance = False
+    for performance in performances:
+        if performance['date'] == source_timezone.localize(datetime(2013, 4, 27)) and performance['track'] == 'Hobart':
+            found_performance = True
+            assert performance['starting_price'] == 22.70
+            break
+
+    assert found_performance