+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [Unreleased]
+
+### Fixed
+
+- Fix handling of valid but mangled performance starting prices (from @justjasongreen)
+
+
 ## [1.0.0b3] - 2016-07-24
 
 ### Changed
@@ -89,6 +96,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Implement DevOps support (from @justjasongreen)
 
 
+[Unreleased]: https://github.com/justjasongreen/punters_client/compare/1.0.0b3...HEAD
 [1.0.0b3]: https://github.com/justjasongreen/punters_client/compare/1.0.0b2...1.0.0b3
 [1.0.0b2]: https://github.com/justjasongreen/punters_client/compare/1.0.0b1...1.0.0b2
 [1.0.0b1]: https://github.com/justjasongreen/punters_client/compare/1.0.0a4...1.0.0b1